Add files via upload
South German Credit Dataset Analysis

Overview
This repository contains code and resources for analyzing the 'South German Credit' dataset. This dataset is an improved version of the older Statlog German credit dataset, with fixed mistakes and additional information, making it more suitable for studying credit risk and making informed decisions regarding creditworthiness.

Purpose

The primary purpose of this project is to:
Analyze the traits and characteristics of the South German Credit dataset.
Develop and test machine learning models to predict credit risk based on the dataset.
Explore ways to enhance the assessment of creditworthiness using advanced analytical techniques.